requirements.txt
discord.py
dropbox
psycopg2-binary